% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newrelic_api_key.R
\name{newrelic_api}
\alias{newrelic_api}
\title{Create API credentials for New Relic}
\usage{
newrelic_api(account_id, nrdb_api_key = NULL, rpm_api_key = NULL)
}
\arguments{
\item{account_id}{the account ID}

\item{nrdb_api_key}{the REST API key from your account settings}

\item{rpm_api_key}{the Insights API key from your insights account settings}
}
\value{
the API credentials object
}
\description{
Create API credentials for New Relic
}

