% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrdb_query}
\alias{nrdb_query}
\title{NRDB Query}
\usage{
nrdb_query(nr_credentials, nrql_query)
}
\arguments{
\item{nr_credentials}{the credentials object created with \code{newrelic_api()}}

\item{nrql_query}{the NRQL query to execute}
}
\value{
a data frame with the results
}
\description{
NRDB Query
}

