% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrdb_session_ids}
\alias{nrdb_session_ids}
\title{Get User Session IDs}
\usage{
nrdb_session_ids(nr_credentials, app_id, from = 10, to = 6, size = 50,
  diverse = T)
}
\arguments{
\item{nr_credentials}{the credentials with the Insights API key.}

\item{app_id}{the application id; you can find the id for your application in the RPM application
by using the ID in the URL when viewing the application.}

\item{from}{number of hours into the past to begin the hunt for sessions}

\item{to}{number of hours into the past to end the hunt for sessions.  The default is six because
if you fetch sessions more recent than that they may still be open and therefore incomplete.}

\item{size}{the maximum number of session ids to fetch}

\item{diverse}{a boolean indicating that the session ids returned will be prioritized according to
how many distinct pages were accessed during the session. This is important if you have sessions from
robots or monitors that you want to ignore.}
}
\value{
a character vector of session ids
}
\description{
Get User Session IDs
}

