% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrdb_sessions}
\alias{nrdb_sessions}
\title{Get the Page Views for a list of session Ids}
\usage{
nrdb_sessions(nr_credentials, session_ids, limit = 750)
}
\arguments{
\item{nr_credentials}{the credentials with the Insights API key.}

\item{session_ids}{the list of session ids obtained using \code{\link{nrdb_session_ids}}}

\item{limit}{the limit on the number of PageViews in a session to this number; if there are
more than this number of pages the session is skipped.  Default limit is 750 and the max limit
is 1000.}
}
\description{
Get the Page Views for a list of session Ids
}

