% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrdb_top_transactions}
\alias{nrdb_top_transactions}
\title{Get the top transactions.}
\usage{
nrdb_top_transactions(account_id, api_key, app_id, limit = 10,
  event_type = "Transaction", end_time)
}
\arguments{
\item{account_id}{your New Relic account ID}

\item{api_key}{your New Relic NRDB (Insights) API key}

\item{app_id}{the application with the transactions}

\item{limit}{to limit the number of transactions returned}

\item{event_type}{the event table to use, either \code{PageView} or (default) \code{Transaction}}
}
\value{
a data frame with two variables, \code{name} and \code{count}
}
\description{
This will return a list of transaction names and their call count for the last 30 minutes
ordered according to call count.  If an end_time is provided then a 30 minute period just
prior to the end_time is examined.
}

