% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrql.timestamp}
\alias{nrql.timestamp}
\title{Generate a formatted, quoted date string suitable for use in SINCE/UNTIL clauses
in NRQL.}
\usage{
nrql.timestamp(ts)
}
\arguments{
\item{ts}{a datetime value as a POSIXt object or a unix Epoch time in seconds}
}
\value{
a character vector with a quoted datetime value
}
\description{
Generate a formatted, quoted date string suitable for use in SINCE/UNTIL clauses
in NRQL.
}
