% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricdata_api.R
\name{rpm_applications}
\alias{rpm_applications}
\title{Query for New Relic Applications.}
\usage{
rpm_applications(account_id, api_key, host = "api.newrelic.com",
  verbose = F)
}
\arguments{
\item{account_id}{your New Relic account ID}

\item{api_key}{your New Relic APM REST API key}

\item{host}{use to proxy requests through an intermediate host you can override the default host
of \code{api.newrelic.com}.}

\item{verbose}{show extra information about the calls}
}
\value{
a data table with observations for each application and variables for the id, name, and throughput
}
\description{
This function returns a list of applications in the given account.
}
\seealso{
\code{\link{newrelic_api}}
}
