% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricdata_api.R
\name{rpm_query}
\alias{rpm_query}
\title{Query New Relic Application Metricdata}
\usage{
rpm_query(nr_credentials, app_id, duration = 60 * 60, end_time = Sys.time(),
  period = NULL, metrics = "HttpDispatcher", verbose = F,
  values = "average_response_time", cache = T)
}
\arguments{
\item{nr_credentials}{the credentials obtained from calling \code{newrelic_api()}`}

\item{app_id}{the application id}

\item{duration}{length of time in seconds of the time window of the metric data}

\item{end_time}{the end of the time window}

\item{period}{duration of each timeslice in seconds, minimum of 60}

\item{metrics}{a character vector of metric names}

\item{verbose}{show extra information about the calls}

\item{values}{a character vector of metric values (calls_per_minute, average_response_time, etc)}

\item{cache}{TRUE if the query results should be cached to local disk}
}
\value{
a data table with observations as timeslices and the timeslice start time, metric name, and values in the variables
}
\description{
This executes a query using the underlying Metric Data REST API at api.newrelic.com.
Depending on the duration and period the data may be fetched in successive batches
due to the limitations on the amount of data that can be retrieved.
}
\seealso{
\code{newrelic_api}
}

