% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{sample_events}
\alias{sample_events}
\title{Retrieve a sample of events for an application.}
\usage{
sample_events(account_id, api_key, app_id, end_time = Sys.time(),
  limit = 100, event_type = "Transaction")
}
\arguments{
\item{account_id}{your New Relic account ID}

\item{api_key}{your New Relic NRDB (Insights) API key}

\item{app_id}{the application with the transactions}

\item{end_time}{the upper bound on the time range for getting events}

\item{limit}{the number of events to retrieve}

\item{event_type}{}
}
\value{
a dataframe with limit rows and the union of all attributes in all sampled events.
}
\description{
This returns a sample of events of event_type of the given length (limit) for the given application.
The events are fetched in chunks going backwards from the end_time until there are a total of limit
events.  There may be gaps in between the chunks.  You aren't guaranteed to get the complete set of
contiguous events in a given time range, hence the idea that you are only getting a "sample."
}

