% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrdb_query}
\alias{nrdb_query}
\title{Execute NRQL queries on NRDB (Insights).}
\usage{
nrdb_query(account_id, api_key, nrql_query, verbose = F, timeout = 10000)
}
\arguments{
\item{account_id}{your New Relic account ID}

\item{api_key}{your New Relic NRDB (Insights) API key}

\item{nrql_query}{the NRQL query to execute}

\item{verbose}{indicates status information to be printed out}

\item{timeout}{the max time in milliseconds to wait for a response (default: 10,000)}
}
\value{
a data frame with the results
}
\description{
This function is a facade on the NRDB REST API.  It allows you to execute queries
and analyze the results in a data frame.
}
\examples{
    nrdb_query(account_id=-1, api_key='your_nrdb_api_license_key_here',
              nrql_query="SELECT count(*) from PageAction facet name")
}
\seealso{
\href{https://docs.newrelic.com/docs/insights/new-relic-insights/adding-querying-data/querying-your-data-remotely}{New Relic NRQL REST API}

\href{https://docs.newrelic.com/docs/insights/new-relic-insights/using-new-relic-query-language/using-nrql}{NRQL Query Reference}

\href{https://docs.newrelic.com/docs/insights/new-relic-insights/adding-querying-data/querying-your-data-remotely#register}{How to get a NRDB API key}
}
