% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrdb_session_ids}
\alias{nrdb_session_ids}
\title{Get User Session IDs.}
\usage{
nrdb_session_ids(account_id, api_key, app_id, from = 48, to = 1,
  size = 50, min_length = 10, max_length = NULL, min_unique = 1,
  event_type = "PageView")
}
\arguments{
\item{account_id}{your New Relic account ID}

\item{api_key}{your New Relic NRDB (Insights) API key}

\item{app_id}{the application id; you can find the id for your application in the RPM application
by using the ID in the URL when viewing the application.}

\item{from}{number of hours into the past to begin the hunt for sessions}

\item{to}{number of hours into the past to end the hunt for sessions.  The default is six because
if you fetch sessions more recent than that they may still be open and therefore incomplete.}

\item{size}{the maximum number of session ids to fetch}

\item{max_length}{applies a filter with a maximum session length in pages}
}
\value{
a character vector of session ids
}
\description{
This looks in the given time range for the top sessions based 
on the number of distinct pages called
during the session.
}

