% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrdb_api.R
\name{nrdb_sessions}
\alias{nrdb_sessions}
\title{Get the user interactions for a list of session Ids, including Page Views and Page Actions.}
\usage{
nrdb_sessions(account_id, api_key, session_ids, app_id = NULL, limit = NULL,
  from = 24 * 3, event_type = "PageView")
}
\arguments{
\item{account_id}{your New Relic account ID}

\item{api_key}{your New Relic NRDB (Insights) API key}

\item{session_ids}{the list of session ids obtained using \code{\link{nrdb_session_ids}}}

\item{limit}{the limit on the number of PageViews in a session to this number; if there are
more than this number of pages the session is skipped.  Default limit is 750 and the max limit
is 1000.}

\item{from}{the starting point in hours past to search for events}

\item{page_views_only}{if true (default) will not get PageActions}
}
\value{
list of data frames for each session that contain a page view in each row
  and the union of all attributes as columns
}
\description{
Get the user interactions for a list of session Ids, including Page Views and Page Actions.
}

